rename files to loggerv2
